+++ packages/sdk-go/sdk_version.gen.go
@@ -2,4 +2,4 @@
 
 package stagehand
 
-const stagehandSDKVersion = "4.0.1"
+const stagehandSDKVersion = "4.0.2"